+++ example_packages/invalid/main.go
@@ -1,0 +1,9 @@
+package invalid
+
+import (
+	"log"
+
+
+funk main() {
+	return e
+)
